Reorganize embedding tests
